+++ argser/consts.py
@@ -1,6 +1,8 @@
-from typing import TypeVar
+from typing import TypeVar, Union, Type
+
 
 Args = TypeVar('Args')
+ArgsObj = Union[Args, Type[Args]]
 
 TRUE_VALUES = {'1', 'true', 't', 'okay', 'ok', 'affirmative', 'yes', 'y', 'totally'}
 FALSE_VALUES = {'0', 'false', 'f', 'no', 'n', 'nope', 'nah'}