+++ argser/parser.py
@@ -2,10 +2,13 @@ import logging
 import re
 import shlex
 from argparse import ArgumentParser, Namespace
-from typing import Any, List, Type
+from functools import partial
+from types import FunctionType
+from typing import Any, List, Type, Tuple, Dict
 
-from argser.consts import Args, SUB_COMMAND_MARK
+from argser.consts import Args, ArgsObj, SUB_COMMAND_MARK
 from argser.display import print_args, stringify
+from argser.exceptions import ArgserException
 from argser.fields import Opt
 from argser.logging import VERBOSE
 from argser.formatters import ColoredHelpFormatter, HelpFormatter
@@ -30,6 +33,7 @@ def _get_fields(cls: Type[Args]):
         for key, value in cls.__dict__.items()
         if not key.startswith('_')
         and not isinstance(value, type)  # skip built-ins and inner classes
+        and not callable(value)
     }
     fields = {
         k: None for k in ann if k not in fields_with_value and not k.startswith('_')
@@ -46,18 +50,43 @@ def _get_fields(cls: Type[Args]):
     return fields
 
 
+def _extract_methods(args_cls: Type[Args]):
+    return {
+        key: value
+        for key, value in args_cls.__dict__.items()
+        if callable(value) and not key.startswith('__')
+    }
+
+
+def _set_factory_from_class_method(
+    args: Args, option: Opt, methods: Dict[str, FunctionType], key: str
+):
+    if isinstance(option.factory, str):
+        if option.factory not in methods:
+            raise ArgserException(f"Couldn't find method {option.factory}.")
+        method = methods[option.factory]
+    elif option.factory is None:
+        method = methods.get(f'read_{key}', None)
+    else:
+        method = None
+    if method:
+        option.factory = partial(method, args)
+
+
 def _read_args(
-    args_cls: Type[Args],
+    args: Args,
     parser_name='root',
     override=False,
     bool_flag=True,
     prefix='--',
     repl=('_', '-'),
-):
-    args = []
+) -> Tuple[Args, List[Opt], Dict[str, tuple]]:
+    options = []
     sub_commands = {}
+    args_cls = args.__class__
     ann = _collect_annotations(args_cls)
     fields = _get_fields(args_cls)
+    methods = _extract_methods(args_cls)
     for key, value in fields.items():  # type: str, Any
         logger.log(VERBOSE, f"reading {key!r}")
         annotation = ann.get(key)
@@ -65,46 +94,53 @@ def _read_args(
 
         if hasattr(value, SUB_COMMAND_MARK):
             sub_commands[key] = _read_args(
-                value.__class__, dest, bool_flag=bool_flag, prefix=prefix, repl=repl
+                value, dest, bool_flag=bool_flag, prefix=prefix, repl=repl
             )
             continue
         if isinstance(value, Opt):
             option = value
-            option.guess_type_and_nargs(annotation)
             if not option.dest:
                 option.set_dest(dest)
         else:
-            default, constructor, help = value, None, None
+            default, factory, help = value, None, None
             if isinstance(value, tuple):
                 if len(value) == 2:
                     default, help = value
                 elif len(value) == 3:
-                    default, constructor, help = value
+                    default, factory, help = value
                 else:
-                    raise ValueError(
+                    raise ArgserException(
                         f"invalid value for {key}. "
                         f"Tuple structure should be: (default, help) or "
-                        f"(default, constructor, help)"
+                        f"(default, factory, help)"
                     )
             option = Opt(
                 dest=dest,
                 default=default,
                 help=help,
-                constructor=constructor,
+                factory=factory,
                 bool_flag=bool_flag,
                 prefix=prefix,
                 repl=repl,
             )
-            option.guess_type_and_nargs(annotation)
 
+        # read factory method
+        _set_factory_from_class_method(args, option, methods, key)
+
+        # override params based on global params
         if override:
             option.bool_flag = bool_flag
             option.prefix = prefix
             option.repl = repl
-        logger.log(VERBOSE, option.__dict__)
-        args.append(option)
+
+        # setup type (and factory if it is still None)
+        option.guess_type_and_nargs(annotation)
+
+        # update class attribute with populated Opt instance
         setattr(args_cls, key, option)
-    return args_cls, args, sub_commands
+        options.append(option)
+        logger.log(VERBOSE, option.__dict__)
+    return args, options, sub_commands
 
 
 def _join_names(*names: str):
@@ -146,9 +182,9 @@ def _make_parser(
 
     sub_parser = parser.add_subparsers(dest=_uwrap(name))
 
-    for sub_name, (args_cls, args, sub_p) in sub_commands.items():
-        p = getattr(args_cls, '__parser', None)
-        parser_kwargs = getattr(args_cls, '__kwargs', {})
+    for sub_name, (args_ins, args, sub_p) in sub_commands.items():
+        p = getattr(args_ins, '__parser', None)
+        parser_kwargs = getattr(args_ins, '__kwargs', {})
         parser_kwargs.setdefault('formatter_class', formatter_class)
 
         p = _make_parser(
@@ -184,14 +220,10 @@ def _set_values(
     for arg in args:
         setattr(res, arg.name, namespace.__dict__.get(arg.dest))
 
-    for name, (args_cls, args, sub_c) in sub_commands.items():
+    for name, (args_ins, args, sub_c) in sub_commands.items():
         # set values only if sub-command was chosen
         if getattr(namespace, _uwrap(parser_name)) == name:
             sub = getattr(res, name)
-            # Reinitialize instance of sub-command so that nullification of attribute
-            # would not touch original class and inner sub-commands. Mostly useful for
-            # tests.
-            sub = sub.__class__()
             setattr(res, name, sub)
             sub_parser_name = _join_names(parser_name, name)
             _set_values(sub_parser_name, sub, namespace, args, sub_c)
@@ -228,11 +260,21 @@ def _make_shortcuts(args: List[Opt]):
 
 def _make_shortcuts_sub_wise(args: List[Opt], sub_commands: dict):
     _make_shortcuts(args)
-    for name, (args_cls, args, sub_p) in sub_commands.items():
+    for name, (args_ins, args, sub_p) in sub_commands.items():
         _make_shortcuts_sub_wise(args, sub_p)
 
 
-def sub_command(args_cls: Type[Args], parser=None, **kwargs) -> Args:
+def _get_args_instance(args: ArgsObj):
+    if isinstance(args, type):
+        args = args()
+    if args.__class__.__str__ is object.__str__:
+        setattr(args.__class__, '__str__', stringify)
+    if args.__class__.__repr__ is object.__repr__:
+        setattr(args.__class__, '__repr__', stringify)
+    return args
+
+
+def sub_command(args_cls: ArgsObj, parser=None, **kwargs) -> Args:
     """
     Add sub-command to the parser.
 
@@ -248,12 +290,11 @@ def sub_command(args_cls: Type[Args], parser=None, **kwargs) -> Args:
     >>> args = parse_args(Data, 'sub -a 2')
     >>> assert args.sub.a == 2
     """
-    setattr(args_cls, '__str__', stringify)
-    setattr(args_cls, '__repr__', stringify)
-    setattr(args_cls, '__parser', parser)
-    setattr(args_cls, '__kwargs', kwargs)
-    setattr(args_cls, SUB_COMMAND_MARK, True)
-    return args_cls()
+    args_ins = _get_args_instance(args_cls)
+    setattr(args_ins, '__parser', parser)
+    setattr(args_ins, '__kwargs', kwargs)
+    setattr(args_ins, SUB_COMMAND_MARK, True)
+    return args_ins
 
 
 def _add_prefixed_key(source: dict, target: dict, prefix: str):
@@ -273,7 +314,7 @@ def _setup_argcomplete(parser, **kwargs):
 
 
 def make_parser(
-    args_cls: Type[Args],
+    args: ArgsObj,
     parser=None,
     make_shortcuts=True,
     bool_flag=True,
@@ -285,9 +326,10 @@ def make_parser(
     **kwargs,
 ):
     """
-    Create arguments parser based on :attr:`args_cls`.
+    Create arguments parser based on :attr:`args`.
 
-    :param args_cls: class with defined arguments
+    :param args: instance of some class with static attributes to be used as
+        ArgParser's options
     :param parser: predefined parser
     :param make_shortcuts: make short version of arguments: ``--abc -> -a``,
         ``--abc_def -> --ad``
@@ -308,33 +350,30 @@ def make_parser(
     :return: instance of ArgumentParser and tuple with options
         (main_options, sub_command_options)
     """
-
-    args_cls, args, sub_commands = _read_args(
-        args_cls, override=override, bool_flag=bool_flag, prefix=prefix, repl=repl
+    args_ins, options, sub_commands = _read_args(
+        args, override=override, bool_flag=bool_flag, prefix=prefix, repl=repl
     )
     if make_shortcuts:
-        _make_shortcuts_sub_wise(args, sub_commands)
+        _make_shortcuts_sub_wise(options, sub_commands)
     # setup parser
     parser_kwargs = parser_kwargs or {}
     _add_prefixed_key(kwargs, parser_kwargs, 'parser_')
     parser_kwargs.setdefault('formatter_class', ColoredHelpFormatter)
-    parser = _make_parser('root', args, sub_commands, parser=parser, **parser_kwargs)
+    parser = _make_parser('root', options, sub_commands, parser=parser, **parser_kwargs)
     # argcomplete
     argcomplete_kwargs = argcomplete_kwargs or {}
     _add_prefixed_key(kwargs, argcomplete_kwargs, 'argcomplete_')
     _setup_argcomplete(parser, **argcomplete_kwargs)
-    return parser, (args, sub_commands)
+    return parser, (options, sub_commands)
 
 
-def populate_holder(
-    parser: ArgumentParser, args_cls: Type[Args], options: tuple, args=None
-):
+def populate_holder(args_ins: Args, parser: ArgumentParser, options: tuple, args=None):
     """
     Parse provided string or command line and populate :attr:`args_cls`
     with parsed values.
 
+    :param args_ins: arguments holder
     :param parser: generated parser
-    :param args_cls: arguments holder
     :param options: tuple with root arguments and sub-command arguments
     :param args: string to parse or ``None``
     :return: instance of :attr:`args_cls` with populated fields.
@@ -344,15 +383,14 @@ def populate_holder(
     namespace = parser.parse_args(args)
     logger.log(VERBOSE, namespace)
 
-    result = sub_command(args_cls)
     args, sub_commands = options
-    _set_values('root', result, namespace, args, sub_commands)
-    setattr(result, '__namespace__', namespace)
-    return result
+    _set_values('root', args_ins, namespace, args, sub_commands)
+    setattr(args_ins, '__namespace__', namespace)
+    return args_ins
 
 
 def parse_args(
-    args_cls: Type[Args],
+    args_cls: ArgsObj,
     args=None,
     *,
     show=None,
@@ -366,7 +404,7 @@ def parse_args(
     Parse arguments from string or command line and return populated
     instance of `args_cls`.
 
-    :param args_cls: class with defined arguments
+    :param args_cls: class with defined arguments or instance of such class
     :param args: arguments to parse. Either string or list of strings or None
         (to read from sys.args)
     :param show:
@@ -393,8 +431,10 @@ def parse_args(
     >>> args = parse_args(Data, '-a 1 -b 2.2 --no-c')
     >>> assert args.a == 1 and args.b == 2.2 and args.c is False
     """
-    parser, options = make_parser(args_cls, **kwargs)
-    result = populate_holder(parser, args_cls, options, args)
+    args_ins = _get_args_instance(args_cls)
+
+    parser, options = make_parser(args_ins, **kwargs)
+    result = populate_holder(args_ins, parser, options, args)
 
     tabulate_kwargs = tabulate_kwargs or {}
     _add_prefixed_key(kwargs, tabulate_kwargs, 'tabulate_')